requirements.txt
# SPDX-FileCopyrightText: 2017 Scott Shawcroft, written for Adafruit Industries
# SPDX-FileCopyrightText: Copyright (c) 2020 Melissa LeBlanc-Williams for Adafruit Industries
#
# SPDX-License-Identifier: MIT

Adafruit-Blinka
adafruit-blinka-displayio
Adafruit-circuitpython-portalbase
adafruit-circuitpython-ili9341
adafruit-circuitpython-stmpe610
adafruit-circuitpython-focaltouch
adafruit-circuitpython-bitmap-font
adafruit-circuitpython-display-text
adafruit-circuitpython-adafruitio
adafruit-circuitpython-neopixel
pillow
wget
